+++ frontend/src/App.jsx
@@ -29,6 +29,7 @@ import EditChapter from './pages/admin/EditChapter';
 import ManagerLogin from './pages/manager/ManagerLogin';
 import ManagerDashboard from './pages/manager/ManagerDashboard';
 import ManagerStories from './pages/manager/ManagerStories';
+import ManagerStoryForm from './pages/manager/ManagerStoryForm';
 import ManagerUsers from './pages/manager/ManagerUsers';
 import ManagerAuthors from './pages/manager/ManagerAuthors';
 import ManagerContacts from './pages/manager/ManagerContacts';
@@ -48,6 +49,8 @@ function App() {
         <Route path="/manager/login" element={<ManagerLogin />} />
         <Route path="/manager/dashboard" element={<ManagerDashboard />} />
         <Route path="/manager/stories" element={<ManagerStories />} />
+        <Route path="/manager/stories/new" element={<ManagerStoryForm />} />
+        <Route path="/manager/stories/:id/edit" element={<ManagerStoryForm />} />
         <Route path="/manager/users" element={<ManagerUsers />} />
         <Route path="/manager/authors" element={<ManagerAuthors />} />
         <Route path="/manager/contacts" element={<ManagerContacts />} />